refactor: coordinate 패키지 position 으로 이름 변경
