+++ backend/src/main/java/kr/touroot/travelplan/dto/response/PlanPositionResponse.java
@@ -1,7 +1,7 @@
 package kr.touroot.travelplan.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.touroot.coordinate.domain.Position;
+import kr.touroot.position.domain.Position;
 import lombok.Builder;
 
 @Builder
